removing mbot.c/h
Working on get rid of mbot.c/h, the existed functions there are mostly related to motors,
thus, if statements checking DIFF DRIVE or OMNI DRIVE everywhere. It might be better
to simply remove mbot.c for now. Also less trouble in dependencies.
