+++ mbot/include/mbot/utils/utils.h
@@ -2,6 +2,8 @@
 #define MBOT_UTILS_H
 #include <mbot/defs/mbot_params.h>
 
+#define UNUSED_DIFF_MOTOR_SLOT       2
+
 int mbot_init_i2c();
 int _mbot_init_i2c(unsigned int pico_sda_pin, unsigned int pico_scl_pin);
 int _check_i2c0_enabled();
