+++ src/mbot_classic.c
@@ -3,14 +3,26 @@
  */
 #include "mbot_classic.h"
 
-// Forward declarations and global variables
+// Global Variables and Forward Declarations
 static bool running = false;
 mbot_params_t params;
 uint64_t timestamp_offset = 0;
 uint64_t global_utime = 0;
 uint64_t global_pico_time = 0;
-bool global_comms_status = COMMS_ERROR; 
+bool global_comms_status = COMMS_ERROR;
 int drive_mode = 0;
+mbot_bhy_config_t mbot_imu_config;
+
+// Forward declaration for internal helper function
+static float _calibrated_pwm_from_vel_cmd(float vel_cmd, int motor_idx);
+
+/*********************************************************************
+ * Main Control Functions
+ * ----------------------------------------------------
+ * These functions are critical for MBot's operation and include the 
+ * main loop and initial setup. Students may be asked to review or 
+ * modify parts of this section depending on their assignment.
+ *********************************************************************/
 
 bool mbot_loop(repeating_timer_t *rt)
 {
@@ -20,21 +32,11 @@ bool mbot_loop(repeating_timer_t *rt)
     mbot_read_imu(&mbot_imu);
     mbot_calculate_motor_vel(mbot_encoders, &mbot_motor_vel);
     
-    if(MBOT_DRIVE_TYPE == DIFFERENTIAL_DRIVE){
-        mbot_calculate_diff_body_vel(   mbot_motor_vel.velocity[params.mot_left], 
-                                        mbot_motor_vel.velocity[params.mot_right], 
-                                        &mbot_vel
-                                    );
-    }
+    mbot_calculate_diff_body_vel(   mbot_motor_vel.velocity[params.mot_left], 
+                                    mbot_motor_vel.velocity[params.mot_right], 
+                                    &mbot_vel
+                                );
     
-    else if(MBOT_DRIVE_TYPE == OMNI_120_DRIVE){
-        mbot_calculate_omni_body_vel(   mbot_motor_vel.velocity[params.mot_left], 
-                                        mbot_motor_vel.velocity[params.mot_right], 
-                                        mbot_motor_vel.velocity[params.mot_back], 
-                                        &mbot_vel
-                                    );
-    }
-
     mbot_calculate_odometry(mbot_vel, MAIN_LOOP_PERIOD, &mbot_odometry);
     mbot_odometry.utime = global_utime;
     // only run if we've got 2 way communication...
@@ -44,34 +46,17 @@ bool mbot_loop(repeating_timer_t *rt)
             mbot_motor_pwm.utime = global_utime;
             mbot_motor_pwm_cmd.pwm[params.mot_right] = _calibrated_pwm_from_vel_cmd(mbot_motor_vel_cmd.velocity[params.mot_right], params.mot_right);
             mbot_motor_pwm_cmd.pwm[params.mot_left] = _calibrated_pwm_from_vel_cmd(mbot_motor_vel_cmd.velocity[params.mot_left], params.mot_left);
-            if(MBOT_DRIVE_TYPE == OMNI_120_DRIVE){
-                mbot_motor_pwm_cmd.pwm[params.mot_back] = _calibrated_pwm_from_vel_cmd(mbot_motor_vel_cmd.velocity[params.mot_back], params.mot_back);
-            }
         }else if(drive_mode == MODE_MBOT_VEL){
             //TODO: open loop for now - implement closed loop controller
-            if(MBOT_DRIVE_TYPE == OMNI_120_DRIVE){
-                mbot_motor_vel_cmd.velocity[params.mot_left] = (SQRT3 / 2.0 * mbot_vel_cmd.vx - 0.5 * mbot_vel_cmd.vy - params.wheel_base_radius * mbot_vel_cmd.wz) / params.wheel_radius;
-                mbot_motor_vel_cmd.velocity[params.mot_right] = (-SQRT3 / 2.0 * mbot_vel_cmd.vx - 0.5 * mbot_vel_cmd.vy - params.wheel_base_radius * mbot_vel_cmd.wz) / params.wheel_radius;
-                mbot_motor_vel_cmd.velocity[params.mot_back] = (mbot_vel_cmd.vy - params.wheel_base_radius * mbot_vel_cmd.wz) / params.wheel_radius;
-                float vel_left_comp = params.motor_polarity[params.mot_left] * mbot_motor_vel_cmd.velocity[params.mot_left];
-                float vel_right_comp = params.motor_polarity[params.mot_right] * mbot_motor_vel_cmd.velocity[params.mot_right];
-                float vel_back_comp = params.motor_polarity[params.mot_back] * mbot_motor_vel_cmd.velocity[params.mot_back];
-                
-                mbot_motor_pwm.utime = global_utime;
-                mbot_motor_pwm_cmd.pwm[params.mot_right] = _calibrated_pwm_from_vel_cmd(vel_right_comp, params.mot_right);
-                mbot_motor_pwm_cmd.pwm[params.mot_back] = _calibrated_pwm_from_vel_cmd(vel_back_comp, params.mot_back);
-                mbot_motor_pwm_cmd.pwm[params.mot_left] = _calibrated_pwm_from_vel_cmd(vel_left_comp, params.mot_left);
-            }else if(MBOT_DRIVE_TYPE == DIFFERENTIAL_DRIVE){
-                mbot_motor_vel_cmd.velocity[params.mot_left] = (mbot_vel_cmd.vx - DIFF_BASE_RADIUS * mbot_vel_cmd.wz) / DIFF_WHEEL_RADIUS;
-                mbot_motor_vel_cmd.velocity[params.mot_right] = (-mbot_vel_cmd.vx - DIFF_BASE_RADIUS * mbot_vel_cmd.wz) / DIFF_WHEEL_RADIUS;
-                
-                float vel_left_comp = params.motor_polarity[params.mot_left] * mbot_motor_vel_cmd.velocity[params.mot_left];
-                float vel_right_comp = params.motor_polarity[params.mot_right] * mbot_motor_vel_cmd.velocity[params.mot_right];
+            mbot_motor_vel_cmd.velocity[params.mot_left] = (mbot_vel_cmd.vx - DIFF_BASE_RADIUS * mbot_vel_cmd.wz) / DIFF_WHEEL_RADIUS;
+            mbot_motor_vel_cmd.velocity[params.mot_right] = (-mbot_vel_cmd.vx - DIFF_BASE_RADIUS * mbot_vel_cmd.wz) / DIFF_WHEEL_RADIUS;
+            
+            float vel_left_comp = params.motor_polarity[params.mot_left] * mbot_motor_vel_cmd.velocity[params.mot_left];
+            float vel_right_comp = params.motor_polarity[params.mot_right] * mbot_motor_vel_cmd.velocity[params.mot_right];
 
-                mbot_motor_pwm.utime = global_utime;
-                mbot_motor_pwm_cmd.pwm[params.mot_right] = _calibrated_pwm_from_vel_cmd(vel_right_comp, params.mot_right);
-                mbot_motor_pwm_cmd.pwm[params.mot_left] = _calibrated_pwm_from_vel_cmd(vel_left_comp, params.mot_left);
-            }
+            mbot_motor_pwm.utime = global_utime;
+            mbot_motor_pwm_cmd.pwm[params.mot_right] = _calibrated_pwm_from_vel_cmd(vel_right_comp, params.mot_right);
+            mbot_motor_pwm_cmd.pwm[params.mot_left] = _calibrated_pwm_from_vel_cmd(vel_left_comp, params.mot_left);
         }else {
             drive_mode = MODE_MOTOR_PWM;
             mbot_motor_pwm.utime = global_utime;
@@ -82,11 +67,6 @@ bool mbot_loop(repeating_timer_t *rt)
         mbot_motor_pwm.pwm[params.mot_right] = mbot_motor_pwm_cmd.pwm[params.mot_right];
         mbot_motor_set_duty(params.mot_left, mbot_motor_pwm_cmd.pwm[params.mot_left]);
         mbot_motor_pwm.pwm[params.mot_left] = mbot_motor_pwm_cmd.pwm[params.mot_left];
-
-        if(MBOT_DRIVE_TYPE == OMNI_120_DRIVE){
-            mbot_motor_set_duty(params.mot_back, mbot_motor_pwm_cmd.pwm[params.mot_back]);
-            mbot_motor_pwm.pwm[params.mot_back] = mbot_motor_pwm_cmd.pwm[params.mot_back];
-        }
 
         // write the encoders to serial
         comms_write_topic(MBOT_ENCODERS, &mbot_encoders);
@@ -106,16 +86,12 @@ bool mbot_loop(repeating_timer_t *rt)
     uint64_t timeout = to_us_since_boot(get_absolute_time()) - global_pico_time;
     if(timeout > MBOT_TIMEOUT_US){
         mbot_motor_set_duty(DIFF_MOTOR_LEFT_SLOT, 0.0);
-        if(MBOT_DRIVE_TYPE == OMNI_120_DRIVE){
-            mbot_motor_set_duty(1, 0.0);
-        }
         mbot_motor_set_duty(DIFF_MOTOR_RIGHT_SLOT, 0.0);
         global_comms_status = COMMS_ERROR;
     }
 
     return true;
 }
-
 
 int main()
 {   
@@ -159,4 +135,48 @@ int main()
         sleep_ms(200);  
         counter++;
     }
+}
+
+
+/******************************************************
+ * Helper Functions
+ * ----------------------------------------------------
+ * These functions are used internally by the main control functions.
+ * They are not intended for modification by students. These functions
+ * provide lower-level control and utility support.
+ ******************************************************/
+int mbot_init_hardware(void){
+    sleep_ms(1000);
+    // Initialize Motors
+    printf("initializinging motors...\n");
+    mbot_motor_init(DIFF_MOTOR_LEFT_SLOT);
+    mbot_motor_init(DIFF_MOTOR_RIGHT_SLOT);
+    printf("initializinging encoders...\n");
+    mbot_encoder_init();
+
+    // Initialize LED
+    printf("Starting heartbeat LED...\n");
+    gpio_init(LED_PIN);
+    gpio_set_dir(LED_PIN, GPIO_OUT);
+
+    mbot_imu_config = mbot_imu_default_config();
+    mbot_imu_config.sample_rate = 200;
+    // Initialize the IMU using the Digital Motion Processor
+    printf("Initializing IMU...\n");
+    mbot_imu_init(&mbot_imu_data, mbot_imu_config);
+    mbot_init_fram();
+    return MBOT_OK;
+}
+
+// Use slope + intercept from calibration to generate a PWM command.
+static float _calibrated_pwm_from_vel_cmd(float vel_cmd, int motor_idx){
+    if (vel_cmd > 0.0)
+    {
+        return (vel_cmd * params.slope_pos[motor_idx]) + params.itrcpt_pos[motor_idx];
+    }
+    else if (vel_cmd < 0.0)
+    {
+        return (vel_cmd * params.slope_neg[motor_idx]) + params.itrcpt_neg[motor_idx];
+    }
+    return 0.0;
 }
